requirements.txt
flask
pytest
requests
python-dotenv
types-requests